requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
numpy
python-multipart